+++ src/host/mod.rs
@@ -243,8 +243,8 @@ pub(crate) use error_emit::try_emit_error;
     target_os = "dragonfly",
     target_os = "freebsd",
     target_os = "netbsd",
-    target_os = "windows",
     target_vendor = "apple",
+    all(windows, any(feature = "asio", feature = "jack")),
     all(
         target_arch = "wasm32",
         target_os = "unknown",
@@ -254,6 +254,7 @@ pub(crate) use error_emit::try_emit_error;
     ),
 ))]
 #[inline]
+#[allow(dead_code)]
 pub(crate) fn frames_to_duration(
     frames: crate::FrameCount,
     rate: crate::SampleRate,